compose/requirements.txt
numpy
pandas
paho-mqtt
matplotlib